pyproject.toml
[project]
name = "acm2"
version = "2.0.0"
description = "API Cost Multiplier 2.0 - Research evaluation and cost tracking platform"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
keywords = ["llm", "evaluation", "research", "gpt-researcher", "cost-tracking"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core Framework
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.1",
    "pydantic-settings>=2.1.0",
    
    # Database
    "sqlalchemy>=2.0.28",
    "alembic>=1.13.0",
    "aiosqlite>=0.19.0",
    
    # GPT-Researcher (main research engine)
    "gpt-researcher>=0.14.0",
    
    # LLM Providers
    "openai>=1.3.3",
    "anthropic>=0.18.0",
    "google-generativeai>=0.4.0",
    
    # Async & HTTP
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
    "websockets>=12.0",
    
    # Config & Environment
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    
    # Utilities
    "tiktoken>=0.7.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "python-multipart>=0.0.6",
    
    # Evaluation
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
acm2 = "app.cli.main:app"

[project.urls]
Homepage = "https://github.com/youruser/acm2"
Documentation = "https://github.com/youruser/acm2#readme"
Repository = "https://github.com/youruser/acm2"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
include = ["app/static/**"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_ignores = true
